2nd add

+++ ft_memmove.c
@@ -6,7 +6,7 @@
 /*   By: mdoroana <mdoroana@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/16 17:27:31 by mdoroana          #+#    #+#             */
-/*   Updated: 2022/02/18 15:18:24 by mdoroana         ###   ########.fr       */
+/*   Updated: 2022/02/22 17:34:56 by mdoroana         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,13 +14,13 @@
 
 void	*ft_memmove(void *dst, const void *src, size_t len)
 {
-	size_t	i;
+	int	i;
 
 	if (!dst || !src)
 		return (NULL);
 	if (dst > src)
 	{
-		i = len - 1;
+		i = (int)len - 1;
 		while (i >= 0)
 		{
 			((unsigned char *)dst)[i] = ((unsigned char *)src)[i];
@@ -30,7 +30,7 @@ void	*ft_memmove(void *dst, const void *src, size_t len)
 	else
 	{
 		i = 0;
-		while (i < len)
+		while (i < (int)len)
 		{
 			((unsigned char *)dst)[i] = ((unsigned char *)src)[i];
 			i++;
